requirements.txt
PyNUTClient == 2.8.2
python-dotenv == 1.0.1
